+++ gimie/sources/__init__.py
@@ -18,11 +18,9 @@
 from abc import ABC, abstractmethod
 from typing import Any, Dict, Optional, Type
 
-from gimie.sources.local import GitExtractor
-from gimie.sources.remote import (
-    GithubExtractor,
-    GitlabExtractor,
-)
+from gimie.sources.git import GitExtractor
+from gimie.sources.github import GithubExtractor
+from gimie.sources.gitlab import GitlabExtractor
 from rdflib import Graph
 
 REMOTE_SOURCES: Dict[str, Any] = {
